Update get_history_by_timestamp to accept float for limit parameter and cast it to int for history retrieval.

+++ naraninyeo/tools/history.py
@@ -1,4 +1,4 @@
-from typing import Annotated, Literal
+from typing import Annotated, Literal, Union
 from datetime import datetime
 from zoneinfo import ZoneInfo
 from haystack.tools import tool
@@ -9,7 +9,7 @@ from naraninyeo.handlers.history import get_history
 def get_history_by_timestamp(
     room_id: Annotated[str, "The room id to get the history from"],
     timestamp: Annotated[str, "The timestamp to get the history from. (YYYY-MM-DD HH:MM:SS) KST"],
-    limit: Annotated[int, "The number of messages to get. (default: 10)"] = 10
+    limit: Annotated[float, "The number of messages to get. (default: 10)"] = 10.0
 ) -> str:
     """
     Retrieves message history from a specified room based on a given timestamp.
@@ -22,7 +22,7 @@ def get_history_by_timestamp(
     
     # Convert KST timestamp string to datetime with KST timezone
     dt = datetime.strptime(timestamp, "%Y-%m-%d %H:%M:%S").replace(tzinfo=ZoneInfo("Asia/Seoul"))
-    history = asyncio.run(get_history(room_id, dt, limit))
+    history = asyncio.run(get_history(room_id, dt, int(limit)))
 
     result = []
     for message in history:
